requirements.txt
altair==5.0.1
attrs==23.1.0
backports.zoneinfo==0.2.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cycler==0.11.0
decorator==5.1.1
Flask==2.2.5
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.5.3
mdurl==0.1.2
numpy==1.21.6
opencv-python==4.7.0.72
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
protobuf==4.24.3
psutil==5.9.5
pyarrow==12.0.1
pydeck==0.8.1b0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.31.0
rich==13.5.2
scipy==1.7.3
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
streamlit==1.23.1
tenacity==8.2.3
thop==0.1.1.post2209072238
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3.1
ultralytics==8.0.114
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
Werkzeug==2.2.3
zipp==3.15.0
